jc 14 re-rework
