+++ src/main/java/homework/romanivanov/javacore/jc14hw/Commodity/Commodity.java
@@ -1,12 +1,19 @@
 package homework.romanivanov.javacore.jc14hw.Commodity;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.Scanner;
+
 public class Commodity {
     private String name;
     private int length;
     private int width;
     private int weight;
+    Scanner sc = new Scanner(System.in);
+    ArrayList<Commodity> commodities;
 
     public Commodity() {
+        commodities = new ArrayList<>();
     }
 
     public Commodity(String name, int length, int width, int weight) {
@@ -14,6 +21,7 @@ public class Commodity {
         this.length = length;
         this.width = width;
         this.weight = weight;
+        commodities = new ArrayList<>();
     }
 
     @Override
@@ -61,4 +69,103 @@ public class Commodity {
         this.weight = weight;
         return this;
     }
+    public void addCommodity(){
+            Commodity commodity = new Commodity();
+            System.out.println("Введіть назву товара");
+            commodity.setName(sc.next());
+            System.out.println("Введіть довжину товара");
+            commodity.setLength(sc.nextInt());
+            System.out.println("Введіть ширину товара");
+            commodity.setWidth(sc.nextInt());
+            System.out.println("Введіть вагу товара");
+            commodity.setWeight(sc.nextInt());
+            commodities.add(commodity);
+        System.out.println("Товар додано");
+        }
+    public void removeByName() {
+        Iterator<Commodity> iterator = commodities.iterator();
+        System.out.println("Введіть ім'я товара яке хочете видалити");
+        String name = sc.next();
+        while (iterator.hasNext()) {
+            if (iterator.next().getName().equalsIgnoreCase(name)) {
+                iterator.remove();
+                System.out.println("Товар видалено");
+            }
+        }
+    }
+    public void sortByLength() {
+        commodities.sort(new CommodityLengthComparator());
+        for (Commodity c : commodities) {
+            System.out.println(c);
+        }
+
+    }
+
+    public void sortByWidth() {
+        commodities.sort(new CommodityWidthComparator());
+        for (Commodity c : commodities) {
+            System.out.println(c);
+        }
+    }
+
+    public void sortByName() {
+        commodities.sort(new CommodityNameComparator());
+        for (Commodity c : commodities) {
+            System.out.println(c);
+        }
+    }
+
+    public void sortByLWeight() {
+        commodities.sort(new CommodityWeightComparator());
+        for (Commodity c : commodities) {
+            System.out.println(c);
+        }
+    }
+
+    public void findByIndex() {
+        boolean b = true;
+        System.out.println("Введіть i-тий елемент колекції колеції");
+        int index = sc.nextInt();
+        for (int i = 0; i < commodities.size(); i++) {
+            if (commodities.get(i) == commodities.get(index)) {
+                System.out.println(commodities.get(i));
+                b = false;
+
+            }
+        }
+        if(b){
+            System.out.println("Такого елемента не існує");
+        }
+    }
+
+
+    public void changeCommodity() {
+        boolean b = true;
+        System.out.println("Введіть ім'я товара яке хочете замінити");
+        String name = sc.next();
+        for (Commodity c : commodities) {
+            if (c.getName().equalsIgnoreCase(name)) {
+                commodities.set(commodities.indexOf(c), getCommodity());
+                System.out.println("Товар змінено");
+                b = false;
+            }
+        }
+        if (b){
+            System.out.println("Такого тавара не знайдено");
+        }
+    }
+
+    public Commodity getCommodity() {
+        Commodity commodity = new Commodity();
+        System.out.println("Введіть назву товара");
+        commodity.setName(sc.next());
+        System.out.println("Введіть довжину товара");
+        commodity.setLength(sc.nextInt());
+        System.out.println("Введіть ширину товара");
+        commodity.setWidth(sc.nextInt());
+        System.out.println("Введіть вагу товара");
+        commodity.setWeight(sc.nextInt());
+        return commodity;
+    }
+
 }
